+++ app/scripts/loadout/dimLoadoutPopup.directive.js
@@ -35,6 +35,12 @@
         '    <div class="loadout-set">',
         '      <span class="button-name button-full" ng-click="vm.itemLevelingLoadout($event)"><i class="fa fa-level-up"></i> Item Leveling</span>',
         '    </div>',
+        '    <div class="loadout-set">',
+        '      <span class="button-name button-full" ng-click="vm.gatherEngramsLoadout($event)"><img class="fa" src="/images/engram.svg" height="12" width="12"/> Gather Engrams</span>',
+        '    </div>',
+        '    <div class="loadout-set">',
+        '      <span class="button-name button-full" ng-click="vm.startFarmingEngrams($event)"><img class="fa" src="/images/engram.svg" height="12" width="12"/> Engrams to Vault</span>',
+        '    </div>',
         '    <div class="loadout-set" ng-if="vm.previousLoadout">',
         '      <span class="button-name button-full" ng-click="vm.applyLoadout(vm.previousLoadout, $event)"><i class="fa fa-undo"></i> {{vm.previousLoadout.name}}</span>',
         '    </div>',
@@ -44,9 +50,9 @@
     };
   }
 
-  LoadoutPopupCtrl.$inject = ['$rootScope', 'ngDialog', 'dimLoadoutService', 'dimItemService', 'dimItemTier'];
+  LoadoutPopupCtrl.$inject = ['$rootScope', 'ngDialog', 'dimLoadoutService', 'dimItemService', 'dimItemTier', 'toaster', 'dimEngramFarmingService'];
 
-  function LoadoutPopupCtrl($rootScope, ngDialog, dimLoadoutService, dimItemService, dimItemTier) {
+  function LoadoutPopupCtrl($rootScope, ngDialog, dimLoadoutService, dimItemService, dimItemTier, toaster, dimEngramFarmingService) {
     var vm = this;
     vm.previousLoadout = dimLoadoutService.previousLoadouts[vm.store.id];
 
@@ -126,6 +132,7 @@
 
     vm.applyLoadout = function applyLoadout(loadout, $event) {
       ngDialog.closeAll();
+      dimEngramFarmingService.stop();
 
       if (loadout === vm.previousLoadout) {
         vm.previousLoadout = undefined;
@@ -138,7 +145,7 @@
     };
 
     // A dynamic loadout set up to level weapons and armor
-    vm.itemLevelingLoadout = function minBlueLoadout($event) {
+    vm.itemLevelingLoadout = function itemLevelingLoadout($event) {
       var applicableItems = _.select(dimItemService.getItems(), function(i) {
         return i.canBeEquippedBy(vm.store) &&
           i.talentGrid && !i.talentGrid.xpComplete; // Still need XP
@@ -237,6 +244,45 @@
 
       var loadout = optimalLoadout(applicableItems, bestItemFn, 'Maximize Light');
       vm.applyLoadout(loadout, $event);
+    };
+
+    // A dynamic loadout set up to level weapons and armor
+    vm.gatherEngramsLoadout = function gatherEngramsLoadout($event) {
+      var engrams = _.select(dimItemService.getItems(), function(i) {
+        return i.isEngram() && i.sort !== 'Postmaster';
+      });
+
+      if (engrams.length === 0) {
+        toaster.pop('warning', 'Gather Engrams', 'No engrams are available to transfer.');
+        return;
+      }
+
+      var itemsByType = _.mapObject(_.groupBy(engrams, 'type'), function(items, type) {
+        // No more than 9 engrams of a type
+        return _.first(items, 9);
+      });
+
+      // Copy the items and mark them equipped and put them in arrays, so they look like a loadout
+      var finalItems = {};
+      _.each(itemsByType, function(items, type) {
+        if (items) {
+          finalItems[type.toLowerCase()] = items.map(function(i) {
+            return angular.copy(i);
+          });
+        }
+      });
+
+      var loadout = {
+        classType: -1,
+        name: 'Gather Engrams',
+        items: finalItems
+      };
+      vm.applyLoadout(loadout, $event);
+    };
+
+    vm.startFarmingEngrams = function startFarmingEngrams($event) {
+      ngDialog.closeAll();
+      dimEngramFarmingService.start(vm.store);
     };
 
     // Generate an optimized loadout based on a filtered set of items and a value function